+++ kliko/schemas/validators.py
@@ -1,9 +1,0 @@
-
-print("GIJS")
-
-
-def type_validator(value, rule_obj, path):
-    print(value)
-    print(rule_obj)
-    print(path)
-    return True